pyproject.toml
[tool.poetry]
name = "clinical-pipeline"
version = "0.1.0"
description = ""
authors = ["Eugene Tsuprun"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
transformers = "^4.47.1"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
pytorch-pretrained-bert = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
